Add color log output for debug, warnings, and errors.
This makes use of the third-party colorlog module to add color to some
of our log types. We only do this when outputting directly to a
terminal.

Testing Done:
Tested the different log types when outputting to a terminal and when
outputting to a log file or `less`. Saw the color only in the terminal
case.

Reviewed at https://reviews.reviewboard.org/r/7716/