Add utilities for finding channels, releases and rules in packages.
This adds a handful of utility functions that make it easy to get data
on packages that match certain criteria, such as rules that match a
version and the current system, or the current release channel.

Unit tests were added for each of these cases.

Testing Done:
Unit tests pass.

Used these in a custom command that tested this functionality.

Reviewed at https://reviews.reviewboard.org/r/7725/