Improve package data loading and error handling.
When looking up a package from the repository, we now pass the base path
in, helping to ensure the relative URLs are all correct.

We also listen for errors in the lookup, and represent that with a more
useful exception.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/7733/