Add a class for handling pending installs of packages and dependencies.
PendingInstall is used to queue up one or more packages for
installation. It handles checking if the packages can be installed on
the system, and handles fetching package bundles for dependencies for
those packages (and for dependencies of dependencies). It also provides
the caller with the order in which packages must be installed.

By using PendingInstall, the caller can be sure they're getting
everything they need to install the requested set of packages, and that
they'll be installed in the order required.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/7734/